+++ src/HardwareAbstractionLayer/hal_timerInterrupt.cpp
@@ -6,10 +6,10 @@
 #include "pd_timerInterrupt.h"
 #endif
 
-static uint32_t slot0time = 0;
-static uint32_t slot1time = 0;
-static uint32_t slot2time = 0;
-static uint32_t slot3time = 0;
+static float slot0time = 0.0f;
+static float slot1time = 0.0f;
+static float slot2time = 0.0f;
+static float slot3time = 0.0f;
 
 namespace hal {
     void initTimerInterrupt0() {
@@ -53,15 +53,15 @@ namespace hal {
 
     }
 
-    void setSlot0Time(uint32_t usec){slot0time = usec;};
-    void setSlot1Time(uint32_t usec){slot1time = usec;};
-    void setSlot2Time(uint32_t usec){slot2time = usec;};
-    void setSlot3Time(uint32_t usec){slot3time = usec;};
+    void setSlot0Time(float usec){slot0time = usec;};
+    void setSlot1Time(float usec){slot1time = usec;};
+    void setSlot2Time(float usec){slot2time = usec;};
+    void setSlot3Time(float usec){slot3time = usec;};
 
-    uint32_t getSlot0Time(){return slot0time;};
-    uint32_t getSlot1Time(){return slot1time;};
-    uint32_t getSlot2Time(){return slot2time;};
-    uint32_t getSlot3Time(){return slot3time;};
+    float getSlot0Time(){return slot0time;};
+    float getSlot1Time(){return slot1time;};
+    float getSlot2Time(){return slot2time;};
+    float getSlot3Time(){return slot3time;};
 
 
 
@@ -105,8 +105,5 @@ namespace hal {
         return 0;
 #endif
     }
-
-        
-
 
 }
